+++ CHANGELOG.md
@@ -8,3 +8,7 @@
 	* Created errors module for Scrapedia's custom errors;
 	* Started creating unit tests for the class Scrapedia;
 	* Added cache tools to Scrapedia;
+	* Segmented requesting methods (\_\_full_teams, \_\_full_championships) from accessible methods (teams, championships) to use the cache internally at Scrapedia instances;
+	* Added ScrapediaChampionship class to access data related to specific championships;
+
+* 06-17-2019: